Inherit from people controller
