+++ app/controllers/applications/children/information_controller.rb
@@ -1,43 +1,8 @@
 module Applications
   module Children
-    class InformationController < MainController
+    class InformationController < Applications::People::InformationController
       include Incredible::Wizard
-      expose :application
-      expose :child, -> { application.children.find(params[:child_id]) }
-      
-      prepend_before_action :update_child
-
-      form 'children'
-      
-      def show
-        render_wizard nil, template: "children/#{template}"
-      end
-      
-      def update
-        render_wizard child
-      end
-      
-      def new
-        redirect_to wizard_path(steps.first, child_id: child.id)
-      end
-      
-      private
-      
-      def update_child
-        child.update_attributes(permitted_params) if params[:person]
-      end
-      
-      def permitted_params
-        params.require(:person).permit(
-          :first_name,
-          :last_name,
-          :gender,
-          :date_of_birth,
-          :relationship,
-          :school,
-          :school_contact
-        )
-      end
+      expose :person, -> { application.children.find(params[:child_id]) }
     end
   end
 end
