+++ app/controllers/applications/children_elsewhere/information_controller.rb
@@ -1,8 +1,8 @@
 module Applications
   module ChildrenElsewhere
-    class InformationController < Applications::Children::InformationController
-      include Incredible::Wizard      
-      expose :child, -> { application.children_elsewhere.find(params[:children_elsewhere_id]) }
+    class InformationController < Applications::People::InformationController
+      include Incredible::Wizard
+      expose :person, -> { application.children_elsewhere.find(params[:children_elsewhere_id]) }
     end
   end
 end
