+++ app/views/applications/children/information/_title_with_name.html.haml
@@ -11,4 +11,4 @@
   /   = question['notice']
 
 %h2
-  = question['title'] % { name: child.first_name, number: application.children_count }
+  = question['title'] % { name: person.first_name, number: application.children_count }
