--- app/views/children/_title_with_name.html.haml
@@ -1,14 +1,0 @@
-%section.section
-  %h1
-    = question['heading']
-  %p
-    = question['intro']
-
-  %p
-    = question['extra']
-
-  / %p.box
-  /   = question['notice']
-
-%h2
-  = question['title'] % { name: child.first_name, number: application.send(:count_method) }
