+++ app/views/people/questions.html.haml
@@ -1,4 +1,4 @@
-= simple_form_for child, url: wizard_path, method: :put do |f|
+= simple_form_for person, url: wizard_path, method: :put do |f|
   - @questions.each do |q|
     = render partial: q['widget'].to_s, locals: { form: f, question: q }
   %section.actions
