--- config/forms/children_elsewhere.yml
@@ -1,68 +1,0 @@
-form:
-  name: 'Children'
-  template: :check_if_you_can_foster
-  steps:
-    name:
-      template: :questions
-      questions:
-        - name: :section
-          heading: Tell us about who lives in your home
-          widget: :title
-        - name: :sub_section
-          title: What is the %{number} child's full name?
-          widget: :title_with_name
-        - name: :first_name
-          title: First name(s)
-          widget: :input
-        - name: :last_name
-          title: Last name
-          widget: :input
-    gender:
-      template: :questions
-      questions:
-        - name: :section
-          heading: Tell us about who lives in your home
-          widget: :title
-        - name: :sub_section
-          title: What is %{name} gender?
-          widget: :title_with_name
-        - name: :gender
-          widget: :radio
-    dob:
-      template: :questions
-      questions:
-        - name: :section
-          heading: Tell us about who lives in your home
-          widget: :title
-        - name: :sub_section
-          title: What is %{name}'s date of birth?
-          widget: :title_with_name
-        - name: :date_of_birth
-          widget: :child_dob
-    relationship:
-      template: :questions
-      questions:
-        - name: :section
-          heading: Tell us about who lives in your home
-          widget: :title
-        - name: :sub_section
-          title: What is %{name}'s relationship to you?
-          widget: :title_with_name
-        - name: :relationship
-          title: Enter childs relationship to you
-          widget: :select
-    school:
-      template: :questions
-      questions:
-        - name: :section
-          heading: Tell us about who lives in your home
-          widget: :title
-        - name: :sub_section
-          title: Where does %{name} go to school?
-          widget: :title_with_name
-        - name: :school
-          title: Name of the school
-          widget: :input
-        - name: :school_contact
-          title: School phone number
-          widget: :input
